Organise the code better.
There is now a top-level include directory containing the public
headers, in a process_sandbox/ subdirectory, which can be installed
as-is.

Source files and private headers are in src/.

The shared.h header, which was left over from the old smoke test and
whose include was copied and pasted everywhere by mistake is now gone.
