--- experiments/process_sandbox/shared.h
@@ -1,6 +1,0 @@
-// Copyright Microsoft and Project Verona Contributors.
-// SPDX-License-Identifier: MIT
-
-#include <zlib.h>
-int sum(int, int);
-int crash();
