+++ experiments/process_sandbox/tests/sandbox-fake-open.cc
@@ -1,10 +1,9 @@
 // Copyright Microsoft and Project Verona Contributors.
 // SPDX-License-Identifier: MIT
 
-#include "cxxapi/cxxsandbox.h"
-#include "filetree.h"
-#include "sandbox.hh"
-#include "shared.h"
+#include "process_sandbox/cxxsandbox.h"
+#include "process_sandbox/filetree.h"
+#include "process_sandbox/sandbox.h"
 
 #include <stdio.h>
