+++ experiments/process_sandbox/tests/sandbox-zlib.cc
@@ -1,11 +1,11 @@
 // Copyright Microsoft and Project Verona Contributors.
 // SPDX-License-Identifier: MIT
 
-#include "cxxapi/cxxsandbox.h"
-#include "sandbox.hh"
-#include "shared.h"
+#include "process_sandbox/cxxsandbox.h"
+#include "process_sandbox/sandbox.h"
 
 #include <stdio.h>
+#include <zlib.h>
 
 using namespace sandbox;
