+++ experiments/process_sandbox/tests/sandboxlib-fake-open.cc
@@ -1,10 +1,9 @@
 // Copyright Microsoft and Project Verona Contributors.
 // SPDX-License-Identifier: MIT
 
-#include "cxxapi/cxxsandbox.h"
-#include "platform/platform.h"
-#include "sandbox.hh"
-#include "shared.h"
+#include "process_sandbox/cxxsandbox.h"
+#include "process_sandbox/platform/platform.h"
+#include "process_sandbox/sandbox.h"
 
 #include <sys/syscall.h>
 #include <unistd.h>
