+++ experiments/process_sandbox/tests/sandboxlib-upcall-recursive.cc
@@ -1,9 +1,8 @@
 // Copyright Microsoft and Project Verona Contributors.
 // SPDX-License-Identifier: MIT
 
-#include "cxxapi/cxxsandbox.h"
-#include "sandbox.hh"
-#include "shared.h"
+#include "process_sandbox/cxxsandbox.h"
+#include "process_sandbox/sandbox.h"
 
 int test(int idx, int count)
 {
