+++ experiments/process_sandbox/tests/sandboxlib-zlib.cc
@@ -1,9 +1,9 @@
 // Copyright Microsoft and Project Verona Contributors.
 // SPDX-License-Identifier: MIT
 
-#include "cxxapi/cxxsandbox.h"
-#include "sandbox.hh"
-#include "shared.h"
+#include "process_sandbox/cxxsandbox.h"
+#include "process_sandbox/sandbox.h"
+#include <zlib.h>
 
 extern "C" void sandbox_init()
 {
